+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Fix missing key login_socks5_login
 - Fix creating contacts from email address in message
 - Fix two different about dialogs on macOS #2280
+- Fix error in calculatePageKey
+- Fix special messages not getting fetched on incoming messages
+- Fix react adjacent warning in info message
 
 ### Changed
 - exclude more unused files from installation package